+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - `icon-position` for `mj-social-element`
   - pass the `border` attribute from `mj-social` to its children
   - `title` (`mj-button`), `sizes`/directional border attributes (`mj-image`), `role` (`mj-table`) and `font-weight` (`mj-accordion-title`)
+  - `mj-carousel`: `thumbnails="supported"`
 - support `mj-include` with `type="css"` and `type="html"`
 - support `mj-include` of files which contain an `<mj-head>`
 - support `<mj-class>` with child elements (default attributes for nested elements)